feat: surface merge conflict warning in checks panel
