+++ src/renderer/src/store/slices/github.test.ts
@@ -36,6 +36,7 @@ function makePR(overrides: Partial<PRInfo> = {}): PRInfo {
     url: 'https://example.com/pr/12',
     checksStatus: 'pending',
     updatedAt: '2026-03-28T00:00:00Z',
+    mergeable: 'UNKNOWN',
     ...overrides
   }
 }
